+++ core/src/main/kotlin/net/corda/core/flows/NotaryWireFormat.kt
@@ -1,14 +1,12 @@
 package net.corda.core.flows
 
 import net.corda.core.contracts.StateRef
-import net.corda.core.crypto.*
-import net.corda.core.identity.Party
+import net.corda.core.crypto.DigitalSignature
+import net.corda.core.crypto.SecureHash
+import net.corda.core.crypto.TransactionSignature
 import net.corda.core.serialization.CordaSerializable
-import net.corda.core.serialization.serialize
 import net.corda.core.transactions.CoreTransaction
 import net.corda.core.transactions.SignedTransaction
-import java.security.InvalidKeyException
-import java.security.SignatureException
 
 /**
  * A notarisation request specifies a list of states to consume and the id of the consuming transaction. Its primary
@@ -36,34 +34,6 @@ class NotarisationRequest(statesToConsume: List<StateRef>, val transactionId: Se
 
     /** States this request specifies to be consumed. Sorted to ensure the serialized form does not get affected by the state order. */
     val statesToConsume: List<StateRef> get() = _statesToConsumeSorted // Getter required for AMQP serialization
-
-    /** Verifies the signature against this notarisation request. Checks that the signature is issued by the right party. */
-    fun verifySignature(requestSignature: NotarisationRequestSignature, intendedSigner: Party) {
-        val signature = requestSignature.digitalSignature
-        if (intendedSigner.owningKey != signature.by) {
-            val errorMessage = "Expected a signature by ${intendedSigner.owningKey}, but received by ${signature.by}}"
-            throw NotaryInternalException(NotaryError.RequestSignatureInvalid(IllegalArgumentException(errorMessage)))
-        }
-        // TODO: if requestSignature was generated over an old version of NotarisationRequest, we need to be able to
-        // reserialize it in that version to get the exact same bytes. Modify the serialization logic once that's
-        // available.
-        val expectedSignedBytes = this.serialize().bytes
-        verifyCorrectBytesSigned(signature, expectedSignedBytes)
-    }
-
-    private fun verifyCorrectBytesSigned(signature: DigitalSignature.WithKey, bytes: ByteArray) {
-        try {
-            signature.verify(bytes)
-        } catch (e: Exception) {
-            when (e) {
-                is InvalidKeyException, is SignatureException -> {
-                    val error = NotaryError.RequestSignatureInvalid(e)
-                    throw NotaryInternalException(error)
-                }
-                else -> throw e
-            }
-        }
-    }
 }
 
 /**
