+++ core/src/main/kotlin/net/corda/core/internal/notary/UniquenessProvider.kt
@@ -1,0 +1,24 @@
+package net.corda.core.internal.notary
+
+import net.corda.core.contracts.StateRef
+import net.corda.core.contracts.TimeWindow
+import net.corda.core.crypto.SecureHash
+import net.corda.core.flows.NotarisationRequestSignature
+import net.corda.core.identity.Party
+
+/**
+ * A service that records input states of the given transaction and provides conflict information
+ * if any of the inputs have already been used in another transaction.
+ *
+ * A uniqueness provider is expected to be used from within the context of a flow.
+ */
+interface UniquenessProvider {
+    /** Commits all input states of the given transaction. */
+    fun commit(
+            states: List<StateRef>,
+            txId: SecureHash,
+            callerIdentity: Party,
+            requestSignature: NotarisationRequestSignature,
+            timeWindow: TimeWindow? = null
+    )
+}
