+++ docs/source/changelog.rst
@@ -6,12 +6,16 @@ release, see :doc:`upgrade-notes`.
 
 Unreleased
 ==========
-
+  
 * Refactor RPC Client Kryo observable serialiser into it's own sub module
 
 * Fix CORDA-1403 where a property of a class that implemented a generic interface could not be deserialised in
   a factory without a serialiser as the subtype check for the class instance failed. Fix is to compare the raw
   type.
+  
+* Due to ongoing work the experimental interfaces for defining custom notary services have been moved to the internal package.
+  CorDapps implementing custom notary services will need to be updated, see ``samples/notary-demo`` for an example.
+  Further changes may be required in the future.
   
 * Fixed incorrect exception handling in ``NodeVaultService._query()``.
