+++ node/src/main/kotlin/net/corda/node/services/transactions/RaftNonValidatingNotaryService.kt
@@ -1,9 +1,9 @@
 package net.corda.node.services.transactions
 
 import net.corda.core.flows.FlowSession
-import net.corda.core.flows.NotaryFlow
+import net.corda.core.internal.notary.NotaryServiceFlow
+import net.corda.core.internal.notary.TrustedAuthorityNotaryService
 import net.corda.core.node.ServiceHub
-import net.corda.core.node.services.TrustedAuthorityNotaryService
 import java.security.PublicKey
 
 /** A non-validating notary service operated by a group of mutually trusting parties, uses the Raft algorithm to achieve consensus. */
@@ -12,7 +12,7 @@ class RaftNonValidatingNotaryService(
         override val notaryIdentityKey: PublicKey,
         override val uniquenessProvider: RaftUniquenessProvider
 ) : TrustedAuthorityNotaryService() {
-    override fun createServiceFlow(otherPartySession: FlowSession): NotaryFlow.Service {
+    override fun createServiceFlow(otherPartySession: FlowSession): NotaryServiceFlow {
         return NonValidatingNotaryFlow(otherPartySession, this)
     }
